move project generation to starter api helper
